+++ pywr/optimisation/moea.py
@@ -15,7 +15,12 @@ class InspyredOptimisationModel(Model):
         variables = []
         variable_map = [0, ]
         for var in self.variables:
-            variable_map.append(variable_map[-1]+var.size)
+            size = var.double_size + var.integer_size
+
+            if size <= 0:
+                raise ValueError('Variable parameter "{}" does not have a size > 0.'.format(var.name))
+
+            variable_map.append(variable_map[-1]+size)
             variables.append(var)
 
         self._variables = variables
@@ -38,9 +43,25 @@ class InspyredOptimisationModel(Model):
 
         values = []
         for var in self._variables:
-            l, u = var.lower_bounds(), var.upper_bounds()
-            for i in range(var.size):
-                values.append(random.uniform(l[i], u[i]))
+            lower = []
+            upper = []
+            if var.double_size > 0:
+                lower.append(var.get_double_lower_bounds())
+                upper.append(var.get_double_upper_bounds())
+
+            if var.integer_size > 0:
+                lower.append(var.get_integer_lower_bounds())
+                upper.append(var.get_integer_upper_bounds())
+
+            lower = np.concatenate(lower)
+            upper = np.concatenate(upper)
+
+            if len(lower) != len(upper):
+                raise ValueError('Upper and lower bounds are different lengths. Malformed bound data from Parameter:'
+                                 ' "{}"'.format(var.name))
+
+            for i in range(len(lower)):
+                values.append(random.uniform(lower[i], upper[i]))
 
         return values
 
@@ -48,8 +69,16 @@ class InspyredOptimisationModel(Model):
         fitness = []
         for i, candidate in enumerate(candidates):
             for ivar, var in enumerate(self._variables):
-                j = slice(self._variable_map[ivar], self._variable_map[ivar+1])
-                var.update(np.array(candidate[j]))
+                l, u = self._variable_map[ivar], self._variable_map[ivar+1]
+                if var.double_size > 0:
+                    dj = slice(l, l + var.double_size)
+                    var.set_double_variables(np.array(candidate[dj]))
+
+                if var.integer_size > 0:
+                    ij = slice(l + var.double_size, u)
+                    # Convert integer variables to ints
+                    ints = np.rint(np.array(candidate[ij])).astype(np.int)
+                    var.set_integer_variables(ints)
 
             self.reset()
             self.run()
@@ -59,8 +88,19 @@ class InspyredOptimisationModel(Model):
 
     def bounder(self, candidate, args):
         for ivar, var in enumerate(self._variables):
-            lower = var.lower_bounds()
-            upper = var.upper_bounds()
+            lower = []
+            upper = []
+            if var.double_size > 0:
+                lower.append(var.get_double_lower_bounds())
+                upper.append(var.get_double_upper_bounds())
+
+            if var.integer_size > 0:
+                lower.append(var.get_integer_lower_bounds())
+                upper.append(var.get_integer_upper_bounds())
+
+            lower = np.concatenate(lower)
+            upper = np.concatenate(upper)
+
             j = slice(self._variable_map[ivar], self._variable_map[ivar+1])
             candidate[j] = np.minimum(upper, np.maximum(lower, candidate[j]))
         return candidate